@@ -21,11 +21,12 @@
     "owl.carousel": "^2.3.4",
     "popper.js": "^1.16.1",
     "query-string": "^7.1.1",
-    "react": "^18.0.0",
+    "react": "^18.1.0",
     "react-chartjs-2": "^4.1.0",
-    "react-dom": "^18.0.0",
+    "react-dom": "^18.1.0",
     "react-ga": "^3.3.0",
     "react-i18next": "^11.16.6",
+    "react-loader-spinner": "^6.0.0-0",
     "react-owl-carousel": "^2.3.3",
     "react-owl-carousel2": "^0.3.0",
     "react-paginate": "^8.1.3",
@@ -33,10 +34,11 @@
     "react-redux": "^8.0.1",
     "react-router": "^5.2.0",
     "react-router-dom": "^5.2.0",
-    "react-scripts": "5.0.1",
+    "react-scripts": "^5.0.1",
     "react-select": "^5.3.0",
     "reactstrap": "^9.0.2",
     "redux": "^4.2.0",
+    "semantic-ui-react": "^2.1.2",
     "shards-react": "^1.0.3",
     "shards-ui": "^3.0.0",
     "validator": "^13.7.0",
