@@ -9,15 +9,18 @@ import "animate.css";
 import { LanguageProvider } from "./Service/LanguageContext";
 import { BrowserRouter } from "react-router-dom";
 import App from "./App";
-ReactDOM.render(
+import { createRoot } from "react-dom/client";
+
+const container = document.getElementById("root");
+const root = createRoot(container);
+root.render(
   <Provider store={configureStore()}>
     <BrowserRouter>
       <LanguageProvider>
         <App />
       </LanguageProvider>
     </BrowserRouter>
-  </Provider>,
-  document.getElementById("root")
+  </Provider>
 );
 // If you want to start measuring performance in your app, pass a function
 // to log results (for example: reportWebVitals(console.log))
