@@ -1,9 +1,16 @@
-function validateEmail(email) {
+export const validateEmail = (email) => {
   const re =
     /^(([^<>()[\]\\.,;:\s@"]+(\.[^<>()[\]\\.,;:\s@"]+)*)|(".+"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$/;
   return re.test(String(email).toLowerCase());
-}
-const CheckPhoneNumber = async (number) => {
+};
+export const CheckPhoneNumber = (number) => {
   var Regex = /\b[\+]?[(]?[0-9]{2,6}[)]?[-\s\.]?[-\s\/\.0-9]{3,15}\b/m;
   return Regex.test(number);
+};
+export const CheckInputs = (values, error) => {
+  var result = {};
+  for (const property in values)
+    if (!values[property] && !error[property])
+      result[property] = "Required input";
+  return result;
 };
