@@ -1,2 +1,2 @@
 export const ApiBaseUrl2 = "https://localhost:44318/";
-export const ApiBaseUrl = "https://api.edf.ly/";
+export const ApiBaseUrl = "http://127.0.0.1:8000/";
